Created an updated breakout PCB for the TSL1401CL sensor
Used EAGLE CAD to design a breakout board PCB for the TSL1401CL sensor.
Eight of the breakout boards are being fabricated on the LPKF protomat
PCB mill.

The breakout boards will be used to develop several 2-axis filament
sensor prototypes over the next few weeks.

The breakout board pinout can be found in the datasheet for the
TSL1401CL sensor. The EAGLE .brd and .sch files can be found in the
'eagle' directory in the Github repository.

Each breakout board contains two holes which can be used for mounting
purposes if needed. Pads for a decoupling capacitor are located as
closely as possible to the TSL1401CL sensor.