Added firmware support for the STM32F3's integrated programmable gain amplifier
Firmware support for the STM32F3 programmable gain amplifier has been
added.

The programmable gain amplifier can be configured in non-inverting
configuration with a selectable gain of 2, 4, 8, or 16.

I was not able to get the amplifier to work as a differentiator since it
cannot be configured in an inverting configuration.

Right now the code has commented out references to the amplifier. It
could potentially be used as a form of output later in this project.