Added lots of documentation and datasheets, soldered PCBs
Designed TSL1401Cl breakout boards in EAGLE and milled eight of them in
the project lab PCB mill. Soldered five of the boards and four of them
are working, one appears to have a short circuit possibly caused by
inaccessible solder paste or excessive heat when soldering.

I wrote the pinout on the breakout boards, but the pins can also be
found in the TSL1401CL datasheet.

I added a wealth of documentation pertaining the the STM32F3 and
TSL1401CL sensor in a new folder on the Github repository. The STM32
documentation is useful because it gives an overview of the STM32F3
peripherals, and there are examples on how to initialize peripherals
like the fast ADC, DAC, and integrated opamp. There are additional code
samples for peripherals online with a simple google search.