Created a two-axis filament diameter sensor simulator in C#
Developed a graphical simulator in C# to visualize the two-axis edge
detection sensor.

[Instructions]
The LEDs, sensors, and filament can be moved by dragging and dropping
the shapes. To constrain movement to be vertical, hold the 'Z' key when
dragging. To constrain movement to be horizontal, hold the 'X' key.

One of the simulator's most useful features is the ability to execute
arbitrary edge detection code through reflection.  The user's edge
detection code is dynamically compiled and executed at runtime . The
result is displayed on screen.

A coordinate reference for the simulator can be seen in the
"coordinate-references.png" file.