Significant rework of the simulator
Made significant changes to the shadow handling in the 2-axis sensor
simulator.

The previous method for calculating the shadow location was to compute
the slope between the led and the filament and then draw a line. The
line was then extended from the filament to the sensor's surface. This
was inaccurate in part because it did not select the correct point on
the filament, but also because the slope calculations should have been
inverted for one of the axes. This caused a bit of floating point error
in one of the axes. The short version is that the old way of handing
shadows was complex and had errors.

I've now deleted the old shadow code and replaced it with a ray casting
algorithm. Rays are cast from the LEDs and are projected onto the
sensor. A shadow is drawn if the ray intersects the filament. The rising
and falling edges of the shadow are then stored in memory.

The real-time code execution was updated to use the new shadow
coordinate values. Unlike the previous shadow implementation, this
implementation is subject to discretization error (much like the real
system is). Unlike the real system, however, there is no way to obtain
sub-pixel resolution here, so the error jumps up and down as the
filament is moved.