Added 8 reference documents for the STM32F3 code
Added several reference documents and pin diagrams for the STM32F3 ADC
peripheral. The diagrams are very useful since the GPIO pins of the
STM32 are mapped to 16 multiplexed ADC input channels and are referenced
in the ADC API driver supplied by ST. The reference documents also show
how to configure ADC1 and ADC2 in Regular Simultaneous Mode for
simultaneous measurements.