Added support for dual ADC measurements
The STM32F3 contains two fast ADCs, ADC1 and ADC2, which are tightly
coupled and can be utilized in a master-slave configuration. Both ADCs
can be triggered to start a conversion simultaneously and the conversion
value can be read in parallel.

I've changed the initialization functions for the ADC peripherals so
that they operate in Regular Simultaneous Mode, which will allow us to
read both sensor axes at the same time.

I've also configured the ADCs to operate nearly 10 times faster. I found
several optimizations in the reference documents and updated the code to
make use of them. They should be operating at 5Msps.