Improved heuristic edge detection
Improved the heuristic edge detection of the TSL1401CL code. The code
now refuses to accept edges that lie on the first or last pixel of the
sensor, since that indicates that the true edge most likely lies outside
of the sensing area.

Right now the heuristic edge detection will reject edges that:
> Are unreasonably narrow
> Are unreasonably wide
> Lie on the first or last pixel of the TSL1401CL