Added a new I2C communication module
Added a new module for I2C communication to the sensor code. The only
dependency for this code is the STMicroelectronics STM32F3 standard
peripheral library. The current code also includes a rudimentary
bit-bang implementation of I2C communication for debugging purposes
(commented out).

The I2C module in the standard peripheral library is complex and
requires a thorough understanding of the STM32F3's DMA, NVIC and GPIO
peripherals.

This code will be used to implement I2C communication with the standard
RepRap firmwares, in addition to analog and pwm protocols.