Fixed bugs with the I2C code
Fixed several issues with the I2C code, and ended up porting most of the
code from another I2C file that was used to communicate with the
STM32F3's magnetic compass and gyroscope. The code was tested with a Bus
Pirate and an Arduino, and after debugging issues with the Arduino Wire
library, the I2C code now finally works.