Added centered BoxFilter and FiniteDifference
The edge-detection code now uses CenteredBoxFilter and
CenteredFiniteDifference.

For the CenteredBoxFilter, at any given pixel i, the pixels from (i-N)
to (i+N) are averaged, where N is DET_BOX_FILTER_LENGTH. The first N
pixels equal the value given to the (N+1) pixel; and similarly the last
N pixels are given the same value as the (N+1)th last pixel.

For CenteredFiniteDifferences, the first pixel is given the same value
as the 2nd pixel and the final pixel is given the same value as the 2nd
last pixel.