Added a circular buffer module (RING)
Added a new module with a circular buffer that could potentially be used
to let the sensor handle the filament buffer instead of the printer
firmware (to reduce ram usage). The STM32 has significantly more excess
ram than the Atmega processor. This code should work, but is not
currently used by any other module.