Fixed bugs in both Box Filter functions
In (old) BoxFilter: there used to be a bug where the averaged value for
x[i] was subtracted from the moving sum instead of the old value for
x[i].

In (new) CenteredBoxFilter: the (M+1)th last value was never changed,
where M is the box filter length. This lead to a large difference being
picked up by the difference function, making an edge get detected nearly
always at the edge of the sensor. This has been fixed and tested.

Finally, the divisors used in each box filter was replaced with a
constant integer, that we can change to 1 if we decide to make that
optimization.