Fixed a bug that prevented FIL_LastExtruderPosition_MM from being reset
Fixed a bug that prevented the sensor values from being updated when a
second print was started. Previously, FIL_LastExtruderPosition_MM would
not be reset when a new print was started, meaning that no new
volumetric multipliers would be requested until the extruder position
exceeded the highest extruder position of the previous print. To fix
this, FIL_LastExtruderPosition_MM is now reset when a G92 gcode command
is used to modify the extruder position. This is not a perfect solution
and will likely cause issues with dual extrusion support.